BpskDetector: Re-center wave forms translated along the y axis.
It has been observed that some audio hardware records wave forms
translated along the y axis. Thus, when two cycles of a PSK signal
are combined, they will not sum to zero upon a phase inversion, and
our detection algorithm will fail.

This change recenters to buffered audio sample so that it is centered
around zero. Two unit tests are included to exercise the improvement.